Removed old plant sync and updated it to OxySync
